Pull API documentation out into a top-level section.
This PR seeks to take an initial step in improving our documentation by
refactoring the API documentation.

- Visibility: API documentation was hidden underneath the ngrok Platform
  section making it hard to find.
- Unification: API documentation and API reference documentation were
  separated into two top-level sections.
- Defragmentation: API documentation was unnecessarily separated into
  many separated pages without a clear reason which makes it harder to
  read and search the documentation.
- Succinctness/Clarity: As I made changes to fix the above, I ended
  up trimming the API documentation to remove extraneous words, fix
  typos, adjust header sizes,  and do other small quality-of-life fixes.

As part of this change, I needed an API reference index page to link to
so I created one that talks a little bit about how we generate the API
reference documentation.

I'm intending to make similar improvements for the remainder of the
documentation.

It's possible this may break a few links to some lesser-referenced pages
in our documentation. I'm not sure yet how to solve this problem and
could use guidance here.

As part of this change I also removed some unused files under the
blog and api directories.