Fix for next page links (#348)
I noticed a few pages had broken `next` links where it would be pointing
to the page itself. It looks like this is because of how docusaurus
treats sidebar categories that act as their own link rather than just a
category. I went through all the categories to make sure I got all of
those.

With this change, it will be trickier to maintain these links, but we'll
preserve the nice UX for users clicking on the category name in the
sidebar (1 click vs 2).