Refactor header module documentation.
This changeset refactors the documentation for the request headers and
response headers modules. Header template variables are factored out
into a partial and included into each page.

Including the headers in each page unfortunately doesn't allow you to
see subnavigation for the different sections of the variables. This is a
bug in docusaurus. It may make sense to eventually move them back out
into their own page if we can eventually find a home for them. This is
better for the time being.

The section on automatic headers belongs in a more generic section about
the http reverse proxy behaviors of ngrok's edge because they happen
regardless of whether this module is enabled or not. This section will
eventually be moved when a more appropriate place for that type of
documentation exists.