Refactor OAuth module documentation.
This commit refactors the OAuth module documentation to be consistent
with our new module documentation format. It makes the following
changes:

- Greatly expands documentation with additional information about how
  the module works including comprehensive examples for each
  connectivity method
- Fixes numerous errors about how the module works, mostly because the
  behavior of the module has changed over time
- Moved all of the per-provider information to the integration guides
  section.

One last note: I didn't finish the 'Errors' section. I'd like to come
back to that and get this improved documentation out in front of the
world without blocking on completing that section which requires a lot
of in-depth spelunking time to pull together.