Correctly map custom font weights (#433)
This PR modifies our font declarations to properly map to our font
weights. Currently, we don't map Euclid Semibold to the `600` weight.
This causes browsers to render a faux bold which affects readability.
This PR fixes that by spec'ing those weights to a single `EuclidSquare`
font name, and changing font files per weight.

In the overview, you'll see lots of faux bold fonts get changed to a
real semibold. You'll notice the top menu also got a little thicker,
since it was intended to be the "medium" font weight. Now it's actually
rendered as such.

## Before
![Screenshot 2023-10-05 at 9 43
56 AM](https://github.com/ngrok/ngrok-docs/assets/1369864/22e2dde7-9d3a-412f-bbab-7337a382c10f)

## After
![Screenshot 2023-10-05 at 9 45
15 AM](https://github.com/ngrok/ngrok-docs/assets/1369864/a84f772b-1332-4765-9e77-3a3cbe7d32a0)