use updated method of signing repos
`apt-key` and `/etc/apt/trusted.gpg.d` are deprecated due to security concerns.  the new recommended method is to directly reference the key in the repository declaration.

one possible issue users could run into is if the `/etc/apt/keyrings` directory is not already created, but it's an easy enough fix.

some references for the deprecation and changes:

https://wiki.debian.org/DebianRepository/UseThirdParty
https://www.linuxuprising.com/2021/01/apt-key-is-deprecated-how-to-add.html
https://askubuntu.com/questions/1286545/what-commands-exactly-should-replace-the-deprecated-apt-key?newreg=c7d51f5afac945fea79db7a504d68e76