feat: restructure add-headers, cleanup imports, better codeblock (#842)
Note, this is a draft of the new actions documentation format and file
format.

---

Started re-working action documentation. There is a list below of
actions that have been converted over. Here is a preview of what the
page now looks like on load:


![image](https://github.com/user-attachments/assets/5e733b83-cb3c-47a3-b73c-ebe9630c99a7)

Originally I was going with a generative approach to the documentation
using JSON Schema or a JSON Schema-like document but because of the
behavior details in actions and the way docusaurus is architected, I
gave this up.

Instead, this approach groups action information into a single folder
structure:

- `/traffic-policy/actions/{action_name}/`
  - `index.mdx` (Overview)
  - `behavior.mdx` (Behavior Details)
  - `config.mdx` (Configuration Reference)
  - `examples/index.mdx` (Examples Index)
  - `examples/*.mdx` (Individual Example)
  - `variables.mdx` (Action Result Variables)

When an action has protocol or phase specific details we introduce a
folder and move specifics into it like so:

- `/traffic-policy/actions/deny/tcp/`
  - `examples/index.mdx` (TCP Specific Examples Index)
  - `examples/*.mdx` (TCP Specific Individual Example)
  - `behavior.mdx` (TCP Behavior Details)
  - `config.mdx` (TCP Configuration Reference)

And adjust references to use `/tcp/` instead of the root folder.

While some information _may be duplicated_ in this structure, it allows
all information to be isolated to a single folder which brings us one
step closer to a single root traffic policy documentation section. This
is just the first step towards that.

Lastly, this also brings in other updates:

- Updated examples.
- Updated configuration tables (required fields, minimums and maximums).
- Updated behavior documentation for actions.
- Updated `ConfigExample` component, it now supports showing line
numbers and line highlighting through the metastring.
- Updated `ConfigExample` component, it also now supports getting rid of
the `// snippet` line.
- Clean up the imports we don't need to use `../../` and can use the
root `/docs` or `/src` or `@site/src` if we want to use aliases.
- Added `cross-env` so build works on all machines.

---

Todo:

- [x] Convert actions
- [x] Clean up old files

Actions to be converted:

- [x] add-headers
- [x] compress-response
- [x] custom-response
- [x] deny
- [x] jwt-validation
- [x] log
- [x] rate-limit
- [x] redirect
- [x] remove-headers
- [x] restrict-ips
- [x] url-rewrite

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
Co-authored-by: Cody A Price <cody@ngrok.com>