policy_engine: rename inbound/outbound to new phase names (#955)
Resolves ngrok-private/ngrok#33333

This change updates policy references from `inbound`/`outbound` to the
new phases: `on_tcp_connect`, `on_http_request`, and `on_http_response`.

The one thing I'm unsure about is the Go SDK examples. I updated them to
read from a file to match the other examples, but I'm unsure if there is
a different way we want to present it.

NOTE: this does not include the API examples, as those are auto
generated by a pipeline. I'll do that shortly after this merges.