Streamline+update Auth0 integration guide (#1201)
Suggesting some improvements to this guide based on my own experiences
setting up JWT for #1196, and because this referenced edges and the old
Traffic Policy editor.

I'd like to have a solid and up-to-date guide for reference whenever
we're talking about setting up JWTs with Auth0 instead of repeating the
content.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>