Add tutorial: Deploy a multicloud API gateway (#1196)
This supplants #1161, which I'll close momentarily.

After talking to @shubcodes, we agreed that endpoint pools are not the
right strategy for the most common multicloud requirements. Endpoint
pools could become an *advanced* section of this tutorial, or a
follow-up tutorial.

Also not included here: failover strategy using internal cloud
endpoints. I think the right way is to ship this and write up failover
as a follow-on task later this month.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>