Potentially fix integrations pages (#1211)
I noticed earlier today that all the integrations pages are broken—the
[integrations hub](https://ngrok.com/docs/integrations/) is empty, as
are the individual hubs for each provider.

Error I'm seeing on any integrations page:

```
ZodError: [
  {
    "code": "invalid_type",
    "expected": "string",
    "received": "undefined",
    "path": [
      11,
      "docs",
      0,
      "contentTitle"
    ],
    "message": "Required"
  }
]
```

I poked around a bit and figured that error was getting thrown because
not every integrations guide explicitly has a `title:` in the
frontmatter, and this Zod dude doesn't like an undefined required
string. Making `contentTitle` optional _seems_ to fix the error and
restore the integrations hubs... but at what cost?

---------

Co-authored-by: Shaquil Hansford <liuqahs15@gmail.com>
Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>