Fix self-referential "next" links (#1215)
The next/previous article links at the bottoms of the top-level
"Overview" pages link to themselves.

This is because these overview pages are both the top-level link for the
section (For example, click on Universal Gateway in the nav and it takes
you to the overview), but we also list them as sub-pages under the
section. That means, technically, the next item in the list after the
top-level overview page _is_ the overview page.

```
{
  label: "Universal Gateway",
  // It's the top-level link here:
  link: { type: "doc", id: "universal-gateway/overview" },
  items: [
    // and the first item in the list, here:
    "universal-gateway/overview",
    ...
  ]
}
```

We do this because it's not obvious that you can click the top-level
links with how our current sidebar nav is styled, so it's more
user-friendly to also list the overview page as one of the items.

## Solution

We might be able to swizzle the sidebar component to address this, but
for now we can manually use `pagination_next` in the frontmatter to fix
this (as I've done in this PR)