Update developer productivity guide (#1102)
In starting the work on #1069, we realized that our [developer
productivity
guide](https://ngrok.com/docs/guides/other-guides/security-dev-productivity/)
was likely out of date.

This PR brings the guide up to date with new use cases, instructions
that use Traffic Policy, some new screenshots (and others removed), and
more. This also removes links to the PDF version, since we have no
resourcing in place to update that right now.

A few blockers:

- [x] Need OIDC docs to link to and confirm config
- [x] Need OAuth docs to link to and confirm config
- [ ] ~~New images/diagrams where necessary~~

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
Co-authored-by: Shaquil Hansford <liuqahs15@gmail.com>