Fix lang switcher layout shift (#1258)
There are pages with lots of `<LangSwitcher>` components, like [this
page](https://ngrok.com/docs/traffic-policy/actions/restrict-ips/#restricting-access-with-allow-and-deny-lists),
which uses a ton of `<ConfigExample>`s, which use `<LangSwitcher>` under
the hood. These pages experience layout shift because when you change
the language on one block, it changes the language on all the blocks,
which cumulatively changes the size of the page drastically.

This makes code blocks in the langswitcher collapsible by default, with
a shorter height than standard code blocks, to reduce overall layout
shift. There's probably a better solution but this is good for now.

## Note

This also adds the ability to hide the agent config or traffic policy
tabs from the `<ConfigExample>` component