Update previous API gateway tutorials to bring into parity (v2) (#1238)
Closing #1192 in favor of this PR, as I took a completely different take
on how to organize them after learning a lot from the multicloud one.

This looks like a ton of change, but it's a lot of copy-paste to makes
sure that all three API gateway specific tutorials follow the same
pattern of implementation (as lead by the multicloud tutorial), use the
same demo services, and, for the K8s ones, use the new Operator features
like `AgentEndpoint` resources, instead of Gateway API (which requires
way more complicated YAML configs).

@shubcodes You might be helpful with your experience in these kinds of
setups.
@Alice-Lilith Would love to make sure I'm representing the K8s features
properly—I've tested these to work, but I'm sure there are even better
ways to present the workflows and make it super smooth.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
Co-authored-by: Shaquil Hansford <liuqahs15@gmail.com>