add validate-codeblocks CI linter (#1307)
There are a lot of invalid code blocks in our documentation.

1. Write a script which scans all of the codeblocks in our docs repo and
fails if any of them do not pass a syntax check.
2. Fix all of the problems that the script caught. There are many. Lots
of our k8s yaml manifests had tabs (which are invalid in YAML), lots of
improper indentation, etc.

The actual script of interest is in src/validate-codeblocks/validator.ts

- This script currently only works with JSON/YAML codeblocks. Other
lanauges will need support added for them in the future.
- There are a few codeblocks that we want syntax highlighting on even
though they are intentionally wrong in a few places, for example where
we have ellipses to indicate elided output. In these cases you can now
annotate those code blocks with a `skip-validation` tag and the
codeblock validator will ignore them.