Add redirect and remove old "Using ngrok with Kubernetes" guide (#1316)
Fixes
[#EPD-684](https://linear.app/ngrok/issue/EPD-684/remove-and-redirect-away-from-using-ngrok-with-kubernetes)

This guide is very outdated and should be removed so people don't
accidentally use the old ingress controller or create edges/modules.

I felt like the best option was to point to the base-level K8s docs for
now, but open to suggestions there. And hopefully I did the redirect
script right—it works on my machine™️!