Add new guide for load balancing between K8s services (#1315)
[#CON-135](https://linear.app/ngrok/issue/CON-135/how-to-pool-endpoints-with-k8s-operator)

This is a quick guide that explicitly walks through the process of
setting up `AgentEndpoint`, Ingress, and Gateway API implementations for
Endpoint Pooling.

In writing this I realized how useful a similar walkthrough would be for
the use case at the very end—pooled internal Agent Endpoints in K8s—but
I think that should be a separate guide, as it requires configuring
cloud endpoints and Traffic Policy/`forward-internal`. Open to
suggestions on that, though!

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
Co-authored-by: Shaquil Hansford <liuqahs15@gmail.com>