Add colormode aware image component (#1310)
This PR adds a `<ThemeImage>` component that takes a `darkSrc` and
`lightSrc` so images can change based on the user's browser theme.

## Note

The light mode image will always initially load because docusaurus
doesn't know the colormode on the first render. This issue will cause
dark mode users to see the images show as light for a second, then
switch to dark mode on page loads. This issue will not happen when
swapping between color modes with the color mode switcher near the top
of the page

**Potential fix**: If we implement the mantle color mode switcher, this
problem might not exist. But I don't know for sure. We could also add a
wrapper hook that saves the user's preferred color mode in localstorage
every time it changes


https://github.com/user-attachments/assets/fb53a392-c003-41e4-94fb-76a8f7ae25bf

## Preview here:
-
https://ngrok-docs-git-shaquil-doc-265-add-colormode-a-53c244-ngrok-dev.vercel.app/docs/guides/api-gateway/multicloud#what-youll-learn

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>