Add new guide for load balancing between clouds (#1330)
[#CON-131](https://linear.app/ngrok/issue/CON-131/how-do-i-load-balance-across-clouds)

One last quick guide to push through for using endpoint pools across
clouds. I decided to also bring in the environment agnostic part of this
with the Go SDK, but it definitely adds a tiny bit of bloat—open to
feedback on that part. This then points to using a Cloud Endpoint +
Traffic Policy, which sets them up better to use the mentioned custom
strategies.

Once this and #1322 are through, I'm going to create a new PR that
removes old load balancing guides via Edges and pushes them to one of
these three guides.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
Co-authored-by: Shaquil Hansford <liuqahs15@gmail.com>