Remove unnecessary index pages in guides directory (#1381)
In a few of the guide's sections in the sidebar, we have top-level index
pages that are completely unnecessary. All they do is list the contents
of their directories, which is completely pointless. Here's an example
below.


![image](https://github.com/user-attachments/assets/4ec65dd8-743c-44ce-a736-eddb98f0fd79)

This PR removes those pages and redirects them to a reasonable
alternative.

## How to test

Just open the preview and navigate to the following sections:
- Developer Preview
- Other Guides
- Site-to-Site Security
- Identity-Aware Proxy
- Using ngrok with

In production, clicking these sections will take you to a top-level
page. In the preview, there should be no top-level page.