Eliminate duplicated content and enforce consistent header hierarchy in Traffic Policy Actions docs (#1384)
The Traffic Policy Actions section of the docs takes a "componentized"
approach to the content, which has led to a bug where material is
duplicated on some pages as a result of the pattern being applied
inconsistently. This PR eliminates the duplicated content on the
following pages:

- Add Headers
- Basic Auth
- Circuit Breaker
- Compress Response

It also corrects the header hierarchy by applying the patterns
established on the other pages in this section. While auditing this
section, I noticed that these pages didn't conform to the existing
header patterns:

- Custom Response
- Deny
- Forward Internal
- Set Vars

So I corrected those as well.

In the future, I think we should reconsider this componentized
content—it's cumbersome to edit individual pages and I'm not sure if the
modular benefits are worth the trade-off.