add bindings to ngrok agent config docs (#1429)
We're missing bindings config in the agent config file doc. In general I
think this page could use some TLC or a rewrite, but that's a bit beyond
the scope of the issue I set out to fix. Something we should definitely
look into when we get some time though.

---------

Signed-off-by: Alice-Lilith <a.wasko@ngrok.com>
Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>