Add Minecraft example (#1435)
I felt like it was better to retain the existing Minecraft guide, and
cross-link between them, because the existing Minecraft guide has info
about free plans and troubleshooting that don't necessarily make sense
as part of this specific example.

I couldn't help myself but do a tiny bit of cleanup in the Minecraft
guide while I was in there.