Iterate on Universal Gateway Overview description (#1378)
Following internal discussions on Slack, I'm proposing an update to the
introductory description of Universal Gateway on the Overview page. I
believe this description is more accurate and easier for new users to
understand. (Also corrected repetition of "accelerates.")

While here, I also took a minute to enforce the one-sentence-per-line
style rule and added an Oxford comma toward the bottom of the page.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>