New example: Combine Authentication Methods (#1456)
This example is based on a customer who wants to accept traffic that's
authenticated via IP restrictions `OR` JWT validation. This is a much
more complicated example to explain that most of what's here so far, so
happy to get any feedback on how to make the human readable **What's
happening here** part any clearer.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>