Update Cloud Endpoints Quickstart (#1441)
This PR thoroughly revamps the Cloud Endpoints Quickstart. This doc used
to be deeply nested in the Universal Gateway section but I'm proposing
we relocate it to the Getting Started Section. Ideally I think the user
journey in this section begins with the Agent CLI guide, optionally
proceeds to one of the Agent SDK quickstarts and/or Kubernetes
quickstarts, and then continues to this Cloud Endpoints guide before
setting the user free to explore from there.

Another way to look at it is: Agent CLI quickstart (ngrok 101) + Cloud
Endpoints quickstart (ngrok 201) should arm the new user with all of the
key concepts they need to know to confidently paddle out beyond the
Getting Started shore and into the open ocean of the product docs at
large (300-400-level materials).

This doc encourages the user to get started with cloud endpoints
"clickops" style in the dashboard and then repeat the process with the
API (via the Agent CLI). Along the way they get a deeper look at
endpoint types, bindings, and traffic policies beyond what they learned
in the CLI and SDK quickstarts.

---------

Co-authored-by: Shaquil Hansford <liuqahs15@gmail.com>