FreeBSD Does Not Have 64-bit ARM Agent (#1474)
It's unclear if this is an issue with the docs, or if the engineering
team intends to (but is not) publish a FreeBSD 64-bit ARM artifact of
the agent. In either case, opening this as a simple docs update to raise
awareness.

Note that when downloading the FreeBSD agent, no ARM-64 version is
available in the dropdown: https://ngrok.com/downloads/free-bsd
Similarly, if you manually go to what would be the URL based on the
patterns for all other platforms and architectures, this URL is an error
404:
https://bin.equinox.io/c/bNyj1mQVY4c/ngrok-v3-stable-freebsd-arm64.tgz

As such, this PR simply removes reference to the 64-bit ARM FreeBSD
agent from the docs.