Add example: Multiplex to Internal Services from a Single Domain (#1478)
This PR adds a new example to the gallery and does a few other small
housekeeping things:

- Changes the "CDN" verbiage around Wordpress to clarify that we don't
offer a traditional CDN
- Changes some Traffic Policy structures to fit with grouping best
practices (blog on this coming next week)
- Moves some items in the sidebar around to demote Minecraft

I might take the time in an upcoming PR to fully manage the order in the
sidebar, but out of scope for now.

Also adding @nijikokun + @ishanj12 here as reviewers in case they have
more that we want to talk about RE: multiplexing.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>