Update secrets docs -> open beta + removing 'early access' callouts (#1537)
I'm finding 'Developer Preview' hard to talk about in the blog post
without saying 'Open Beta' to clearly indicate it's open to all.
Therefore, updating this doc to use that same language. I am also
removing the instructions to request Early Access, as we've now enabled
this for all users and this flag is no longer needed, as well as the
disclaimer about supporting a limited swath of fields.