Remove broken markup, remove reference to "clinic" (#1565)
The `:::` syntax doesn't appear to get parsed, and appears in the page
as-is.

<img width="2010" height="464" alt="CleanShot 2025-09-08 at 10 19 41@2x"
src="https://github.com/user-attachments/assets/ef143ed4-5a26-4625-8869-5b1130816d9b"
/>

The word `clinic` in the sentence it is found feels like a typo of some
sort, but I'm not sure what it was originally supposed to be. I've
changed it to "network" as that feels like it works in the sentence.