Update reference.mdx
# ngrok API Reference

We love API reference documentation and we know you do too. As a developer,
it's often the first place to go to understand a system's domain model and to
understand all of its capabilities.

## Correctness

As developers ourselves, we crave confidence in the correctness of our
documentation. To that end, ngrok's API reference documentation is generated
from the same specifications that are used to generate our server-side API
stubs so that you can be confident that the documented parameters match the
implementation.

## Example Correctness

Example requests and responses you see in the API documentation are generated
from a set of integration tests that run against the API so that you can be
confident that the examples in the reference documentation match the behavior
of the API itself.