Install and configure Vale with Google Developer style guide rules (#1621)
this PR installs Vale with (most of) the Google style rules package to
get us started. I made some changes throughout to make it more useful to
us but there's definitely more to be done. This should be enough to
start using Vale right away as a CI check.

I tried getting an LLM to fix the 500ish errors reported but it made a
colossal mess so I had to revert all of those changes. We can start to
actually apply the Vale rules in follow-up PRs.

The command to run it is `pnpm run vale`. There are lots of flags
available to refine it.

Note: once this PR is merged we'll need to turn Vale back on in the
Mintlify dashboard to get the CI checks.

---------

Co-authored-by: Shaquil Hansford <liuqahs15@gmail.com>