Clean up all Vale errors (#1643)
You can test that all of the errors are cleared locally by running `pnpm
run vale --minAlertLevel error`

This PR also downgrades the "Units" rule from error to warning, and we
might want to consider dropping it entirely. I'm not sure it's useful.

It's possible there may be some find+replace typos in here! Hopefully I
caught them all.

I also took the liberty of replacing all variations (that I could find)
of `Step 1:` etc. with `1.` across the board - not technically a Vale
error but I was already editing other similar headers so why not.

There are many instances where "e.g." and "i.e." were used incorrectly
(which is one reason why we should avoid using them). I'm not sure if
it's really worth my time to comb through them one by one to confirm
"for example" and "that is" are always the best replacements—they're not
exactly interchangeable but it's close.