Wrap our own workflow around mintlify's broken link checker
Mintlify's broken link checker doesn't work with its API reference generator.

Because of this, if we have links to the pages mintlify generates for us, the link checker will fail our builds because those pages do not exist in the repo.

Until mintlify implements a solution here, we'll have to do it ourselves.

## What this PR does

This PR adds a workflow that runs mintlify's link-checker, but doesn't fail CI if the links found are part of a list of exceptions (link-exceptions.json).

It also tests if redirects have conflicts, and fails the build if they do (prevents infinite loops)