Propose moving to using our OpenAPI schema. (#1642)
This is in response to an interaction I had on Bluesky over the weekend
(https://bsky.app/profile/stephanietuerk.net/post/3m3l2veuojk2x). Our
current API reference docs don't look and feel like API reference docs.
I noticed Mintlify have support for OpenAPI schemas, and we have an
OpenAPI schema that appears to be up to date, so this PR replaces our
current reference docs with that.

I haven't gone through an updated all of the references to the old API
reference docs yet, and this PR _would_ be a breaking change for anyone
that's linking to our API reference docs from externally. But I really
wanted to get your take on this as an idea. Should we do it? Should we
not? Is there a great reason we _haven't_ done this, or why we opted to
make our own API reference docs _not_ based off of the OpenAPI schema?

---------

Co-authored-by: Shaquil Hansford <liuqahs15@gmail.com>