Fix pnpm fmt command (#1684)
Currently if you run the `pnpm run fmt` command in the package.json file
it attempts to format react code in mdx files and breaks lots of mdx
views.
Seems like we should just ignore mdx files. 

This is only a draft because i wanted to get some opinions here. If this
is fine i can add this as a CI check.

---------

Signed-off-by: Jonathan Stacks <jonstacks@users.noreply.github.com>
Co-authored-by: Jonathan Stacks <jonstacks@users.noreply.github.com>