Actually use generated event-types-table (#1691)
I noticed we weren't actually using the generated _event-types-table.mdx
file anywhere. I assume this wasn't intentional? It seems like it goes
in the observability event logs reference, so that's where I put it.

By generating this page instead of maintaining it manually, we'll always
be up to date.

This PR literally deletes all the content on the /obs/events/reference
page and imports the generated content instead.

Preview:
-
https://ngrok-actually-use-generated-event-types.mintlify.app/obs/events/reference

## Note

If you don't like the layout of the page, we can make edits to the ngrok
go repo to change how the docs are generated. I think we should still
for. now just make sure this page has accurate up to date info by
importing the content.